requirements.txt
#Python >= 2.6
Django>=1.7
django-tastypie>=0.12.1